rabbit_feature_flags: Support mandatory feature flags
Mandatory feature flags are existing feature flags which were introduced
some time ago and we want to make them required and remove the old
compatibility code.

Mandatory feature flags MUST be enabled before upgrading to the RabbitMQ
version where they are marked mandatory. If they are not, RabbitMQ will
refuse to start. The reason is their corresponding compatibility and
migration code is gone.

To make a feature flag mandatory, its `stability` property needs to be
set to `mandatory`.